requirements.txt
faiss-cpu
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
google-generativeai
pypdf
python-dotenv
rank-bm25
jq
